refactor: unify download update selection helpers
